Towards 1.20beta7
String parameters can now be quoted in all G- and M-commands
Added partial support for standstill current reduction (M917)
M0 and M1 no longe turns anything off or run any macros when in
simulation mode

+++ src/RepRap.h
@@ -145,7 +145,7 @@ private:
 	uint32_t fastLoop, slowLoop;
 	uint32_t lastTime;
 
-	uint16_t debug;
+	uint32_t debug;
 	bool stopped;
 	bool active;
 	bool resetting;
